pyproject.toml
[tool.poetry]
name = "yodocus"
version = "0.1.5"
description = "Document Layout Analysis with YOLO"
authors = ["tea-de-kay", ]

[tool.poetry.dependencies]
python = "~=3.12"
onnxruntime = "1.22.*"
opencv-python = "4.12.*"
PyYAML = "6.*"
pillow = "11.*"

huggingface-hub = { version = "0.34.*", optional = true}

[tool.poetry.extras]
huggingface = ["huggingface-hub"]

[tool.poetry.group.dev.dependencies]
pytest = "8.*"
pytest-cov = "6.*"
flake8 = "7.*"
isort = "6.*"
black = "25.*"
pyright = "1.*"


[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests", "tests_integration", "tests_e2e"]
python_files = ["test_*.py", "*Test.py", "*_test.py", ]


[tool.isort]
profile = "black"
line_length = 88
lines_after_imports = 2


[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["ASYNC", "B", "E", "F", "I", "S", "SIM", "UP", "W"]
ignore = ["SIM103", "S101"]

[tool.ruff.lint.pycodestyle]
max-line-length = 120

[tool.ruff.lint.isort]
lines-after-imports = 2


[build-system]
requires = ["poetry-core>=1.0.0", ]
build-backend = "poetry.core.masonry.api"